+++ demo/helper-scripts/demo-pod-with-discovery.yaml.cloverleaf.template
@@ -11,7 +11,7 @@ metadata:
                             {
                                 "matchExpressions": [
                                     {
-                                        "key": "feature.node.kubernetes.io/pstate-turbo",
+                                        "key": "feature.node.kubernetes.io/cpu-pstate.turbo",
                                         "operator": "DoesNotExist"
                                     }
                                 ]
