+++ source/cpu/rdt_amd64.go
@@ -16,7 +16,7 @@ See the License for the specific language governing permissions and
 limitations under the License.
 */
 
-package rdt
+package cpu
 
 import (
 	"sigs.k8s.io/node-feature-discovery/pkg/cpuid"
